Initial work towards parameterizing element drawing ratios.
The element rendering chain functions (Composition.renderMap,
Composition,renderElements, Pics.getTilePic and Pics.baseElementPic) return a
Reader computation, thus allowing the ratios to be propagated from the upper
reaches of the program. The environment is, for now, a Double which sets the
relative road width. It is to be replaced by an ADT with all relevant
parameters.