Shallow cropping of maps to a tile range.
"Shallow" in the sense that the base map is rendered wholly regardless of how
many tiles are requested. Deep cropping (i.e excising the tiles before
rendering) will eventually be implemented. Doing so, however, will probably
call for an overall rethinking of how origins are arranged.

It is tempting to make the bounds displayed to the user one-based rather than
zero-based. Such an adjustment, however, will be deferred until it is clearer
that using two coordinate systems at the same time will not be too confusing.