Improving alignment of segment captions.
Improving, but only to the point at which the output is passable. The caption
placement code will likely be rewritten in The Right Way at some point.

Some adjustments to (hopefully) improve readability of the captions were done
as well. Chances are visible backgrounds for them will have to be added.