Adjustments for compatibility with pre-release threepenny-gui 0.4.0.0 .
Other than eliminating the superfluous reactimate and newEventsTagged in
Util.Reactive.Threepenny, only those changes needed to make the program work
were done. The file writers in the Output module were generalised to MonadIO
so that runRenderMap can be converted to the UI monad in a non-invasive way.

Built on
https://github.com/HeinrichApfelmus/threepenny-gui/tree/3d533790fdd67640ba9cdb86cfd3789c20cbd08a