Packaging flipbook frames with zip-conduit (backend prototype).
The implementation is not at all definitive. Ideally, a port of zip-archive to
io-streams would be used, making it possible to shed a few extra dependencies.
Also, the zip package still has to be made available from the interface. There
is also an unsolved annoyance concering the resulting paths within the zip.