Lenses for TraceAnnotation and TraceOverlays.
Note that liberal sprinkling of lens accessors might, in the wrong
circumstances, lead to noticeable performance degradation. In particular,
there were issues with Annotation.LapTrace, most likely due to the accessors
in the instance implementations. They can be solved in at least three ways:

* Reverting to vanilla field accessors,
* Stop exporting the field names, or
* Change the optimisation level to O2.

We settled for the second option for now, as it is sensible regardless of this
issue; in any case, a move to O2 is worth considering.