Experimenting with bindings to JQuery UI's Autocomplete widget.
The key complication arises from the fact that we need to have the library
fully loaded by the time we call `autocompleteInit` on the text box. As of
Threepenny 0.4.0.2, following usual JavaScript practice by chaining the
initialisation steps with callbacks is not feasible, at least as far as I can
see. Using the `load` event of the `script` element does not work either, as
JQuery (which handles element appendage behind the scenes) will evaluate the
script and eliminate the DOM element. The next best option, then, is to load
the script via a synchronous request. The JavaScript code for doing so was
wrapped with `callFunction` to keep it blocking.

It should be mentioned that the easiest way to load something like JQuery UI
would probably be through `tpCustomHTML`. That, however, would be rather
opaque and inflexible.

See also http://stackoverflow.com/a/2880147/2751851 , from which the sync
request snippet was mostly taken.