Multi-backend support experiments (basic signature adjustments).
The current approach involves explicitly parameterising backends
everywhere there is a `Diagram` or `QDiagram` involved, with the help
of a class of reasonable backends `BeDi`. The place where that has led
to the largest changes is `Annotation.Flipbook`, where many of the
classes have become MPTCs (and similar things will happen in the
`Annotation` module so that font metrics for the various backends can
be handled). There is still one key unsolved issue: how to
programatically change the backend token at the `runRenderMap`
invocation (for instance, based on a command line flag) in a way that
typechecks.

This experiment will stay on hold for now, as the intended target
backend, `Diagrams.Backend.Rasterific`, is currently unusable for our
purposes due to a memory leak of unknown origin.