Using SVGFonts to draw caption annotations.
This is one way of solving the problem with caption backgrounds
introduced in 3dba8c440543 .

The captions currently use the monospace `bit` font, as from the ones
packaged with SVGFonts it is the less thin one. Additionally, it
doesn't take as long to initialise as `lin2`, which has a richer
character set. Bringing in a different font of our choice is something
worth considering.