Adjusting meaning of `Chirality` and removing its `Achiral` value.
With these changes, real chirality (that is, when it actually
corresponds to a difference in the physical layout of the element) and
nominal chirality (that is, when it is assigned to achiral elements
simply as a matter of setting up orientation conventions) were unified.
That means all track elements now have either `Dextral` or `Sinistral`
chiralities, be them real or (in the case of `Dextral`) nominal.
Formerly `Achiral` elements can now be distinguished by the lack of the
`Chiral` property, or by the `NoMirroring` property, which indicates
the chiral counterpart does not exist in Stunts. The goal of the
changes is setting the scene for a chirality and orientation model that
can be used to correctly mirror the representation of a track.

Subsidiary changes:

* Now that left/right corks are `Sinistral`, the base picture in `Pics`
  is the dextral one.

* A `NoRotation` property was introduced for elements, such as trees,
  which only have one orientation. Its role is analogous to that of
  `NoMirroring`.