Upgrading to Threepenny 0.8 (pre-release) to get `loadFile` back.
Also upgrading to Stackage LTS-8.2 and diagrams 1.4, which seemingly
do not require further adjustments.

Cf. https://github.com/HeinrichApfelmus/threepenny-gui/issues/110
and https://github.com/HeinrichApfelmus/threepenny-gui/pull/115 .

The changes here undo the key parts of ee88e45 , which was a workaround
for the lack of `loadFile` in Threepenny 0.6 . Only the minimal
adjustments needed were done here. It may make sense to reconsider some
details, such as the handling of the disambiguating prefixes of
`loadFile`, and whether usin `loadDirectory` instead can eliminate
them.

Note that there seems to be a timing bug involving the Alertify pop-up
that is supposed to show up when flipbook rendering starts. It may be
related to the solution to
https://github.com/HeinrichApfelmus/threepenny-gui/issues/131 , though
for now that is just a conjecture, as I haven't investigated it at all.