Loading the caption font at compile time.
This TH trick doesn't matter all that much in terms of performance, as
the font loading was already being done only once, upon the first
rendering of a caption. Still, it saves time on said first rendering and
replacies an use of unsafePerformIO with something less hairy. Even if
we were to go back to the unsafePerformIO hack, it would be a good idea
to keep that in a separate module.

Note that the TH trick doesn't translate easily to the Pango-based font
handling of Diagrams.Backend.Cairo.Text. We may have to settle for
actually producing diagrams in IO if we decide to use that in the
future.