Using threepenny-gui-flexbox for the UI layout.
Though there are some rough edges in the layout to be dealt with, this
is miles better than whatever there was before already.

I'd really like to use the gap flexbox property, but since it is not
available in Safari yet, and probably will become so in the
short-to-medium term, we'll do without it for the moment.

Util.Threepenny.Flexbox includes a general purpose function to override
only a few flexbox properties, rather than all of them at once.

Note we are using the upstream head of threepenny-gui-flexbox, as the
0.4.0.2 version currently on Hackage won't build with recent GHCs. A
patch fixing that has already been merged. Doing so provided a good
occasion to clean up the main cabal file a bit.