Correcting minor imperfections in the highway transition rendering.
One of them was a simple arithmetic mistake. The other one was a more
subtle issue in which placing polygons of the same colour next to each
other in a specific way appears to make a thin sliver of the background
bleed through the contact line of the polygons when rendering at certain
resolutions.