Support for building the viewer with different diagrams backends.
Backend selection is done through Cabal flags. There are a few details
to be dealt with (for instance, it shouldn't be possible to select PNG
as the output format with the diagrams-svg build). In any case, what we
now have is good enough to try out and compare the backends.

This supersedes the experimental branch at 58d4208. Note that, since
that branch was put on hold, diagrams-rasterific became a viable option
for our purposes. Preliminary testing suggests that, though is slower
than diagrams-cairo for full map rendering, it is faster for flipbook
rendering.