repldump2carto now includes car speed and current gear in its output.
There are many other known fields of CarState that could be included.
Before that, however, it will make sense to either have Cartography to
read the binary repldump output directly or to have repldump2carto
producing a CSV file with all known fields.

Note that there seems to be a discrepancy between the byte counts in the
repldump2carto binary parser and those in the restunts notes. That
should be further investigated, though it's quite possible that I'm just
miscounting it.