Running the main rendering computation in a separate thread.
Said computation used to be ran straight from an event handler. This
additional level of indirection was added in order to avoid keeping
the event network on hold until the main computation concludes. One
way that issue manifests itself is in how the log text area is not
updated until the main computation is concluded.

A significant amount of code around the main action trigger had to be
reworked. While the resulting code is not terribly pretty, it is also
a fair bit more modular than it used to be, which should make it more
amenable to further refactoring.