Add t2c command for exporting parts of trackdata
The command currently supports exporting trackdata09 and trackdata10.
It uses the same format of r2c for exporting, so that the output can
be loaded as if it were a lap trace by the track viewer.

The t2c code was taken from a standalone script originally based on r2c,
so there is a lot of deduplication to be done.