Make CartoT a newtype which uses Writer.CPS
The writer from RWS was the culprit of a space leak in Gallery, and
switching to Writer.CPS from transformers fixes it.

mtl-2.3.1 would have made this fix much easier, as it gives the MTL
instances to RWS.CPS from transformers. There has to be a little
catching up around the ecosystem before it gets comfortable to use
mtl-2.3.1. Even if we choose to keep the CartoT newtype wrapper in the
long term, moving to mtl-2.3.1 would free us from writing a few
instances explicitly.